poc/slack-bot/requirements.txt
slack-bolt>=1.18,<2
slack-sdk>=3.21,<4
python-dotenv>=1.0,<2
